refactor(#133): extract travel listing into travel.js
Moves travel map, lightbox, card/timeline loading, and view-toggle
logic from script.js into a dedicated travel.js module.

- travel.html now loads travel.js instead of script.js
- Travel page visit counter now fires correctly on page load
- script.js is scoped to homepage only (GitHub widget, contact form, home visit counter)
