+++ travel.html
@@ -55,11 +55,11 @@
             <div id="travel-map" class="travel-map" aria-label="Map of travel memories"></div>
 
             <div class="travel-grid" id="travel-grid">
-                <!-- populated from API by script.js -->
+                <!-- populated from API by travel.js -->
             </div>
 
             <div id="travel-timeline" class="timeline">
-                <!-- populated from API by script.js -->
+                <!-- populated from API by travel.js -->
             </div>
 
             <div id="travel-empty" class="travel-empty hidden">
@@ -95,6 +95,6 @@
     <script src="./resources/java/jquery-3.5.1.min.js"></script>
     <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"
             integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo=" crossorigin=""></script>
-    <script type="module" src="./resources/java/script.js"></script>
+    <script type="module" src="./resources/java/travel.js"></script>
 </body>
 </html>
